mocha tests, fix issue with strip comments
